config: add backupDir to config.Create
